Add support for logarithmic scale
This patch adds a new preference choice (linear or logarithmic),
including a corresponding GConf key and applet object attribute. It only
affects the colored bars in the panel, not the graph. The scale slider
specifies the upper bound of the graph, the lower bound is hard-coded to
a 10000th of the upper graph. This allows, for example, to keep an eye
on both local pings in the range of a few ms and remote pings in the
range of seconds.

Conflicts:

	ui/preferences-dialog.ui